pyproject.toml
[tool.poetry]
name = "django_test"
version = "0.1.0"
description = ""
authors = ["Mario Llobet <mllobet@xxxx.xx>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.3.1"
coverage = "^6.4"
flake8 = "^4.0.1"


[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
